+++ src/main/java/svenhjol/charm/feature/chairs/Chairs.java
@@ -18,7 +18,7 @@ import net.minecraft.world.phys.Vec3;
 import svenhjol.charm.Charm;
 import svenhjol.charmony.common.CommonFeature;
 import svenhjol.charmony.feature.advancements.Advancements;
-import svenhjol.charmony_api.event.BlockUseEvent;
+import svenhjol.charmony.api.event.BlockUseEvent;
 
 import java.util.function.Supplier;