+++ src/main/java/svenhjol/charm/feature/vanilla_wood_variants/VanillaWoodVariants.java
@@ -3,7 +3,7 @@ package svenhjol.charm.feature.vanilla_wood_variants;
 import svenhjol.charmony.common.CommonFeature;
 import svenhjol.charmony.enums.VanillaWood;
 import svenhjol.charmony.feature.variant_wood.VariantWood;
-import svenhjol.charmony_api.CharmonyApi;
+import svenhjol.charmony.api.CharmonyApi;
 
 public class VanillaWoodVariants extends CommonFeature {
     @Override