+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 6.0.21
+
+- Increase NearbyWorkstations range to 16 blocks
+- Fix compass overlay compact view coords hidden when at far distance
+- Bump to Charmony 6.15.0
+
 ## 6.0.20
 
 - Fix for unenchanted items being added to middle slot of smithing table